@@ -16,7 +16,7 @@
   "scripts": {
     "test": "",
     "start": "node ./dist/database/src/index.js",
-    "delDist": "ts-node ../utils-build/delDistDir.ts ./database/dist",
+    "delDist": "ts-node ../utils/utils-build/delDistDir.ts ../database/dist",
     "tsc": "tsc --project ./",
     "compile": "npm run delDist && npm run tsc",
     "tslint": "tslint --config ./tslint.json --project ./tsconfig.json",
@@ -27,30 +27,33 @@
   },
   "dependencies": {
     "debug": "^4.1.1",
-    "mongoose": "^5.4.10",
-    "node": "^11.9.0",
+    "mongoose": "^5.4.14",
+    "node": "^11.10.0",
     "winston": "^3.2.1"
   },
   "devDependencies": {
+    "@types/app-root-path": "^1.2.4",
+    "@types/assert": "^1.4.2",
     "@types/chai": "^4.1.7",
-    "@types/debug": "0.0.31",
-    "@types/mocha": "^5.2.5",
-    "@types/mongoose": "^5.3.12",
-    "@types/node": "^10.12.21",
+    "@types/debug": "4.1.1",
+    "@types/mocha": "^5.2.6",
+    "@types/mongoose": "^5.3.17",
+    "@types/node": "^11.9.4",
     "@types/proxyquire": "^1.3.28",
-    "@types/sinon": "^7.0.5",
+    "@types/sinon": "^7.0.6",
     "@types/sinon-chai": "^3.2.2",
+    "@types/winston": "^2.4.4",
     "app-root-path": "^2.1.0",
     "chai": "^4.2.0",
-    "mocha": "^5.2.0",
+    "mocha": "^6.0.0",
     "mocha-sinon": "^2.1.0",
-    "nyc": "^13.2.0",
+    "nyc": "^13.3.0",
     "prettier": "^1.16.4",
     "proxyquire": "^2.1.0",
-    "sinon": "^7.2.3",
+    "sinon": "^7.2.4",
     "sinon-chai": "^3.3.0",
     "ts-node": "^8.0.2",
     "tslint": "^5.12.1",
-    "typescript": "^3.3.1"
+    "typescript": "^3.3.3"
   }
 }
