@@ -1,13 +1,12 @@
 ﻿const modulename: string = __filename.slice(__filename.lastIndexOf('\\'));
-import debugFunction from 'debug';
+import debugFunction = require('debug');
 const debug = debugFunction(`PP_${modulename}`);
 debug(`Starting ${modulename}`);
 
-/* import for types */
-import 'mocha';
-
-/* set up sinon & chai */
+/* set up mocha, sinon & chai */
+import { fail } from 'assert';
 import chai = require('chai');
+import 'mocha';
 import sinon = require('sinon');
 import sinonChai = require('sinon-chai');
 chai.use(sinonChai);
@@ -15,95 +14,108 @@ const expect = chai.expect;
 sinon.assert.expose(chai.assert, {
   prefix: '',
 });
-import { fail } from 'assert';
 
 /* external dependencies */
 import appRootObject = require('app-root-path');
 const appRoot = appRootObject.toString();
-import { Connection } from 'mongoose';
+import { SchemaDefinition } from 'mongoose';
 import path = require('path');
-import proxyquire = require('proxyquire');
+import proxyquireObject = require('proxyquire');
+const proxyquire = proxyquireObject.noPreserveCache();
 
 const indexPath = path.join(appRoot, 'src', 'index');
-// const databasePath = path.join(appRoot, 'src', 'database');
 
-/* import .config and database files */
-import { getMongoUri } from '../src/.config';
-import { Database } from '../src/database';
+/* internal dependencies */
+import {
+  Database,
+  filepaths,
+  getConnectionOptions,
+  getMongoUri,
+} from '../src/.config';
 
 describe('Database connection', () => {
-  /* set up common variables */
+  debug(`Running ${modulename} describe Database.connection`);
+
+  /* set up module  variables */
   let database: Database;
-  let debugStub: any;
-  let debugEnabled = true;
+  const debugEnabled = true;
+  const spyDebug = sinon.spy();
 
-  // /* set up a database schema, collection, and model name */
-  // const testSchema = {
-  //   username: String,
-  //   email: String,
-  // };
-  // const testCollection = 'mochaTest';
-  // const testModel = 'mochaTestModel';
+  /* derive db name from the connection uri */
+  const uri = getMongoUri();
+  const charStart = uri.indexOf('://');
+  const char1 = uri.indexOf('/', charStart + 3); // skip initial :// in uri
+  const char2 = uri.indexOf('?ssl=');
+  const dbName = uri.substring(char1 + 1, char2);
 
-  before('Set up database debug.enabled mock', async () => {
-    debugStub = (prefix: string) => {
+  /* set up a database schema, collection, and model name */
+  const testSchema: SchemaDefinition = {
+    username: String,
+    email: String,
+  };
+  const testCollection = 'mochaTest';
+  const testModel = 'mochaTestModel';
+
+  afterEach('Close database connection', async () => {
+    debug(`Running ${modulename} afterEach Close database connection`);
+
+    /* close dbConnection if open */
+    if (
+      database &&
+      database.dbConnection &&
+      database.dbConnection.readyState === 1
+    ) {
+      await database.closeConnection(database.dbConnection);
+    }
+  });
+
+  it('tests spyDebug', async () => {
+    debug(`Running ${modulename} it tests spyDebug`);
+
+    /* stub index import debugFunction function */
+    const debugFunctionStub = (prefix: string) => {
+      const debugLocal = debugFunction(prefix);
+      debugLocal.enabled = debugEnabled;
       return (message: string) => {
-        debugFunction(prefix).enabled = debugEnabled;
-        debugFunction(prefix)(message);
-        debug(message);
+        spyDebug(message);
+        debugLocal(message);
       };
     };
 
-    /* set up default database */
-    const getDatabase = proxyquire(indexPath, {});
+    /* try connect to database */
+    const getDatabase = proxyquire(indexPath, {
+      debug: debugFunctionStub,
+    });
     database = await getDatabase.runDatabaseApp();
 
-    // debug(modulename + ':\n\n *** Tests start here ***\n\n');
+    if (true) {
+      expect(spyDebug.lastCall.lastArg).to.eql(
+        '\\index.ts: running runDatabaseApp',
+      );
+    }
   });
 
-  afterEach('Close dbConnection & restore database', async () => {
-    /* close dbConnection */
+  it('makes a connection to a database', async () => {
+    debug(`Running ${modulename} it makes a connection to a database`);
+
+    /* close dbConnection if open */
     if (
       database &&
       database.dbConnection &&
       database.dbConnection.readyState === 1
     ) {
       await database.closeConnection(database.dbConnection);
+      expect(
+        database.dbConnection.readyState,
+        'Connection should be closed',
+      ).to.equal(0);
     }
 
-    /* restore default database so not needed within each it */
+    /* connect to database */
     const getDatabase = proxyquire(indexPath, {});
     database = await getDatabase.runDatabaseApp();
-  });
 
-  after('Close dbConnection', async () => {
-    /* close dbConnection */
-    if (
-      database &&
-      database.dbConnection &&
-      database.dbConnection.readyState === 1
-    ) {
-      await database.closeConnection(database.dbConnection);
-    }
-  });
-
-  it('dummy', () => {
-    debugStub('TEST_PREFIX')('TEST_MSG)');
-    if (database && database.dbConnection) {
-      expect(database.dbConnection.readyState).to.eql(1);
-    }
-  });
-
-  it('Connects to a database', async () => {
-    /* get and configure a database */
-
-    /* derive db name from the connection uri */
-    const uri = getMongoUri();
-    const charStart = uri.indexOf('://');
-    const char1 = uri.indexOf('/', charStart + 3); // skip initial :// in uri
-    const char2 = uri.indexOf('?ssl=');
-    const dbName = uri.substring(char1 + 1, char2);
-
+    /* test */
     if (database && database.dbConnection && database.dbConnection.db) {
       expect(
         database.dbConnection.db.databaseName,
@@ -118,31 +130,90 @@ describe('Database connection', () => {
     }
   });
 
-  // it('Connects to an open database', async () => {
-  //   /* test not printing database stats */
-  //   debugEnabled = false;
-  //   database = proxyquire(databasePath, {
-  //     debug: debugStub,
-  //   }).database;
-  //   /* new database instance => configure */
-  //   database.configDatabase(config);
+  it('makes a 2nd connection to a database', async () => {
+    debug(`Running ${modulename} it makes a 2nd connection to a database`);
 
-  //   dbConnection = await database.connectToDB();
-  //   expect(
-  //     dbConnection.db.databaseName,
-  //     'Should return a database connection',
-  //   ).to.equal(config.DB_NAME);
-  //   expect(dbConnection.readyState, 'Connection should be open').to.equal(1);
-  //   await database.closeConnection(dbConnection);
-  // });
+    /* connect to database */
+    const getDatabase = proxyquire(indexPath, {});
+    database = await getDatabase.runDatabaseApp();
 
-  it('Closes an open database connection', async () => {
+    /* test */
+    if (database && database.dbConnection && database.dbConnection.db) {
+      expect(
+        database.dbConnection.db.databaseName,
+        'Should return a database connection named <dbName>',
+      ).to.equal(dbName);
+      expect(
+        database.dbConnection.readyState,
+        'Connection should be open',
+      ).to.equal(1);
+    } else {
+      expect(fail('database or dbConnection or db is falsy'));
+    }
+
+    /* make a 2nd connection to the database */
+    const database2 = await getDatabase.runDatabaseApp();
+
+    /* test */
+    if (database2 && database2.dbConnection && database2.dbConnection.db) {
+      expect(
+        database2.dbConnection.db.databaseName,
+        'Should return a database connection named <dbName>',
+      ).to.equal(dbName);
+      expect(
+        database2.dbConnection.readyState,
+        'Connection should be open',
+      ).to.equal(1);
+
+      /* close 2nd dbConnection */
+      await database2.closeConnection(database2.dbConnection);
+      expect(
+        database2.dbConnection.readyState,
+        'Connection should be closed',
+      ).to.equal(0);
+    } else {
+      expect(fail('database or dbConnection or db is falsy'));
+    }
+  });
+
+  it('fails to connect to a database', async () => {
+    debug(`Running ${modulename} it fails to connect to a database`);
+
+    /* stub the database getMongoUri function */
+    const configStub = {
+      getMongoUri: () => 'dummyUri',
+    };
+
+    /* try connect to database */
+    const getDatabase = proxyquire(indexPath, {
+      './.config': configStub,
+    });
+
+    try {
+      await getDatabase.runDatabaseApp();
+      expect.fail('Should not have reached this point');
+    } catch (err) {
+      expect(err.name, 'Should be a MongoParseError').to.eql('MongoParseError');
+    }
+  });
+
+  it('closes an open database connection', async () => {
+    debug(`Running ${modulename} it closes an open database connection`);
+
+    /* connect to database */
+    const getDatabase = proxyquire(indexPath, {});
+    database = await getDatabase.runDatabaseApp();
+
     if (database && database.dbConnection && database.dbConnection.db) {
       expect(
         database.dbConnection.readyState,
         'Connection should be open',
       ).to.eql(1);
+
+      /* close to database */
       await database.closeConnection(database.dbConnection);
+
+      /* test */
       expect(
         database.dbConnection.readyState,
         'Connection should be closed',
@@ -152,18 +223,21 @@ describe('Database connection', () => {
     }
   });
 
-  it('Closes a closed database connection', async function() {
+  it('closes a closed database connection', async () => {
+    debug(`Running ${modulename} it closes a closed database connection`);
+
     if (database && database.dbConnection && database.dbConnection.db) {
-      expect(
-        database.dbConnection.readyState,
-        'Connection should be open',
-      ).to.eql(1);
+      /* close database */
       await database.closeConnection(database.dbConnection);
       expect(
         database.dbConnection.readyState,
         'Connection should be closed',
       ).to.equal(0);
+
+      /* close again */
       await database.closeConnection(database.dbConnection);
+
+      /* test */
       expect(
         database.dbConnection.readyState,
         'Connection should still be closed',
@@ -173,62 +247,77 @@ describe('Database connection', () => {
     }
   });
 
-  it('Closes an invalid database connection', async function() {
-    // tslint:disable-next-line: no-object-literal-type-assertion
-    const dummy = {} as Connection;
-    const result = await database.closeConnection(dummy);
+  it('closes an invalid database connection', async () => {
+    debug(`Running ${modulename} it closes an invalid database connection`);
+
+    const dummyConnection: any = {};
+    const result = await database.closeConnection(dummyConnection);
     expect(result, 'Returns an error').to.be.instanceOf(Error);
   });
 
-  // it('Fails to connect to a database', async function() {
-  //   /* stub the database server host name function */
-  //   const revertDbHost = config.DB_HOST;
-  //   config.DB_HOST = 'dummyHost';
+  it('Creates a mongoose model', async () => {
+    debug(`Running ${modulename} it creates a mongoose model`);
 
-  //   database = proxyquire(databasePath, {
-  //     [path.join(appRoot, 'dist', '.config')]: config,
-  //   }).database;
-  //   database.configDatabase(config);
+    /* connect to database */
+    const getDatabase = proxyquire(indexPath, {});
+    database = await getDatabase.runDatabaseApp();
 
-  //   try {
-  //     await database.connectToDB();
-  //     expect.fail('Should not have reached this point');
-  //   } catch (err) {
-  //     /* restore database */
-  //     config.DB_HOST = revertDbHost;
+    const model = database.createModel(
+      testModel,
+      testSchema,
+      testCollection,
+      database.dbConnection,
+    );
+    expect(model.collection.name, 'Should return a mongoose model').to.eql(
+      testCollection,
+    );
+  });
 
-  //     expect(err.name, 'Should be a MongoNetworkError').to.eql(
-  //       'MongoNetworkError',
-  //     );
-  //   }
-  // });
+  it('Fails to create a mongoose model', async () => {
+    debug(`Running ${modulename} it fails to create a mongoose model`);
 
-  // it('Creates a mongoose model', async function() {
-  //   /* connect to the database and create a db connection */
-  //   const dbConnection = await database.connectToDB();
-  //   const model = database.createModel(
-  //     testModel,
-  //     testSchema,
-  //     testCollection,
-  //     dbConnection,
-  //   );
-  //   await database.closeConnection(dbConnection);
-  //   expect(model.collection.name, 'Should return a mongoose model').to.eql(
-  //     testCollection,
-  //   );
-  // });
+    /* connect to database */
+    const getDatabase = proxyquire(indexPath, {});
+    database = await getDatabase.runDatabaseApp();
 
-  // it('Fails to create a mongoose model', async function() {
-  //   try {
-  //     database.createModel(
-  //       testModel,
-  //       testSchema,
-  //       testCollection,
-  //       'dummyConnection',
-  //     );
-  //     expect.fail('Should not have reached this point');
-  //   } catch (err) {
-  //     expect(err.name, 'Should be a connection error').to.eql('TypeError');
-  //   }
-  // });
+    const dummyConnection: any = {};
+
+    try {
+      database.createModel(
+        testModel,
+        testSchema,
+        testCollection,
+        dummyConnection,
+      );
+      expect.fail('Should not have reached this point');
+    } catch (err) {
+      expect(err.message, 'Should be a connection error').to.eql(
+        'dbConnection.model is not a function',
+      );
+    }
+  });
+
+  it('tests sending no logger or dumpError', async () => {
+    debug(`Running ${modulename} it tests sending no logger or dumpError`);
+
+    /* connect to database without configuring logger or dumpError */
+    const connectionUrl = getMongoUri();
+    const connectOptions = getConnectionOptions();
+    database = new filepaths.DATABASE.Database(connectionUrl, connectOptions);
+    database.dbConnection = await database.dbConnectionPromise;
+
+    /* test */
+    if (database && database.dbConnection && database.dbConnection.db) {
+      expect(
+        database.dbConnection.db.databaseName,
+        'Should return a database connection named <dbName>',
+      ).to.equal(dbName);
+      expect(
+        database.dbConnection.readyState,
+        'Connection should be open',
+      ).to.equal(1);
+    } else {
+      expect(fail('database or dbConnection or db is falsy'));
+    }
+  });
 });
